Refactoring bundle file extension string usage
